requirements.txt

Django>=3.0.5,<3.1.0
djangorestframework>=3.11.0,<3.12.0
psycopg2>=2.8.5,<2.9.0
Pillow>=7.2.0,<7.3.0

djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1

flake8>=3.7.9,<3.8.0
